Выполнена домащная работа для урока 1
